refactor(settlement): 修改交易类型字段为Integer类型- 将AccountTransactionsPo中的transactionType从Byte改为Integer - 更新AccountTransactionsService接口中的参数类型- 修改AccountTransactionsServiceImpl实现类中的参数类型 - 更新AccountTransactionsVO中的transactionType为Integer类型 - 在OrderSettlementServiceImpl中直接使用整数字面量替换byte强制转换 - 简化订单状态设置方式，去除不必要的byte类型转换

+++ src/main/java/com/studio/settlement/bean/po/AccountTransactionsPo.java
@@ -36,7 +36,7 @@ public class AccountTransactionsPo implements Serializable {
 
             @Schema(description = "交易类型（1:充值 2:消费）")
             @TableField("transaction_type")
-            private Byte transactionType;
+            private Integer transactionType;
 
             @Schema(description = "关联订单号")
             @TableField("related_order_code")
